Move hooks folder inside app
