+++ app/suppliers/[handle]/new/upload-preview-form.tsx
@@ -10,7 +10,7 @@ import { Progress } from '@/components/ui/progress'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import { Textarea } from '@/components/ui/textarea'
 import { Location } from '@/db/constants'
-import { useCreateTile } from '@/hooks/use-create-tile'
+import { useCreateTile } from '@/app/_hooks/use-create-tile'
 import { InsertTileRaw, SupplierRaw, User } from '@/models/types'
 import { enumToPretty } from '@/utils/enum-helpers'
