+++ app/t/[tileId]/page.tsx
@@ -9,7 +9,7 @@ import { Area } from '@/components/ui/area'
 import { Button } from '@/components/ui/button'
 import { Section } from '@/components/ui/section'
 import { Separator } from '@/components/ui/separator'
-import { tileKeys } from '@/hooks/queryKeys'
+import { tileKeys } from '@/app/_hooks/queryKeys'
 import { TileModel } from '@/models/tile'
 import { getAuthenticatedUserId } from '@/utils/auth'
 import { valueToPretty } from '@/utils/enum-helpers'
